Did some stuff, check summary
Made a level transition between levels 3 and 4. Both players step on the cloud and get sent up to level 4.
I tried fixing the buttons but gave up because it broke a little (I reverted the changes)
Made it so you can jump on the cloud sprite in level 4
I added the TitleScreen script to the title screen because it used to be there but I guess it got detached somehow? I dunno.
I reordered all the scenes in the build menu and fixed the level transition script to accommodate that

To Do:
Add victory screen
Fix Level 2 Button puzzle
Add transition between level 2 and 3
make beehive a transition zone to get to the victory screen
Make level 2 actually interesting, I think we should have enough time if we put in a couple hours